Refactor: Extract widgets from file_management and settings features
- Extract FileErrorMessage widget for error state display
- Extract FileEmptyState widget for empty state display
- Extract FileCard widget for individual file display
- Extract FileList widget for file list management
- Extract AgentErrorMessage widget for agent error display
- Extract AgentEmptyState widget for agent empty state
- Clean up file_management_screen.dart (reduced from 347 to 50 lines)
- Improve maintainability and testability across features
- Follow established widget extraction pattern
